refactor: Update middleware imports to use new file structure
