+++ middleware.ts
@@ -1,10 +1,9 @@
 import type { NextRequest } from 'next/server';
-import { CommonResponseFactory } from './app/api/shared/response';
-import { isAuthorized } from './app/api/shared/auth';
+import { CommonResponseFactory } from './app/lib/response';
+import { isAuthorized } from './app/lib/auth';
 import { env } from './app/config/env';
 
 export function middleware(request: NextRequest) {
-  console.warn('middleware is running');
   const actualAuthToken = request.headers.get('Authorization');
   const expectedAuthToken = env.API_TOKEN;
